refactor: organize React app by feature
