+++ web/cypress.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  e2e: {
+    baseUrl: 'http://localhost:5173',
+  },
+};
