+++ web/package.json
@@ -8,6 +8,7 @@
     "build": "vite build",
     "lint": "eslint .",
     "test": "jest",
+    "e2e": "cypress run",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -45,6 +46,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^30.0.4",
     "jest-environment-jsdom": "^30.0.4",
+    "cypress": "^13.6.4",
     "postcss": "^8.5.6",
     "tailwindcss": "^3.4.3",
     "vite": "^7.0.0"
