+++ web/src/App.jsx
@@ -1,9 +1,9 @@
 import React, { Suspense } from "react";
 import AppRoutes from "./routes/AppRoutes";
-import { useAuth } from "./pages/auth/useAuth";
+import { useAuth } from "./features/auth/hooks/useAuth";
 import Loading from "./components/Loading";
 
-const LoginPage = React.lazy(() => import("./pages/auth/LoginPage"));
+const LoginPage = React.lazy(() => import("./features/auth/pages/LoginPage"));
 
 export default function App() {
   const { user } = useAuth();
