+++ web/src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import "./assets/styles/index.css";
 import "sweetalert2/dist/sweetalert2.min.css";
-import { AuthProvider } from "./pages/auth/useAuth.jsx";
+import { AuthProvider } from "./features/auth/hooks/useAuth.jsx";
 import { ThemeProvider } from "./theme/useTheme.jsx";
 import { BrowserRouter } from "react-router-dom";
 import axios from "axios";
